requirements.txt
python-dotenv
langchain-google-genai
google-generativeai
browser-use
asyncio